Move data structures to separate source file.
